+++ app/src/main/java/com/kylecorry/trail_sense/tools/maps/domain/MapCalibration.kt
@@ -3,7 +3,7 @@ package com.kylecorry.trail_sense.tools.maps.domain
 data class MapCalibration(
     val warped: Boolean,
     val rotated: Boolean,
-    val rotation: Int,
+    val rotation: Float,
     val calibrationPoints: List<MapCalibrationPoint>
 ) {
     companion object {
@@ -11,7 +11,7 @@ data class MapCalibration(
             return MapCalibration(
                 warped = false,
                 rotated = false,
-                rotation = 0,
+                rotation = 0f,
                 calibrationPoints = emptyList()
             )
         }
